feat: Add possibility for deferred scheduling
This feature is a first attempt to solve the problems described in #7:
When methods are auto-scheduled, the Executor and ThreadPools are
eagerly set up while the Injector is being created. If an error occurs
during Injector creation those Threads are left dangling in nirvana and
might prevent the JVM from properly shutting down.

Using @ManuallyStarted now marks a scheduled method as deferred. It will
first be scheduled by the time SchedulingService.startManualScheduling()
is called.